backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==1.0.3
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.3
sentence-transformers==2.3.1
huggingface-hub==0.20.3
python-dotenv==1.0.0
openai==1.3.5 